Initial commit
This includes a modification of the Youtuberizer plugin developed during
Devember 2019 as well as some additional code as a stub for being able
to edit the description/title/tags of a video (or come up with new
information to later save).

There is tones more work to do to get this functional.

Clean