Alter the video list data return
In order to be more generally useful, the API endpoint for getting the
contents of a playlist has been modified so that it returns a list of
objects that represent the videos instead of the values that the command
itself would directly use.

This allows it to be used in more situations than just the current 
command.