Enhance the tag syntax
In order to better support some upcoming features, the syntax definition
for tags have been augmented so that everything but a comma can appear
in a tag. This is important since previously all punctuation would be
excluded, but that also disallowed space characters, which can indeed be
part of tags.