Include event handlers for checking fields
This includes event listeners for text views in YouTube detail windows
that verifies that the maximum allowable length is detected and that
anything longer than that is flagged as an error using regions.

This marks title, description and tags that are too long; in the case of
tags the check is done on a tag by tag basis as well as the length of
the field as a whole.