Details are not always required in returns
It transpires that it's possible for fields that YouTube returns to not
exist, such as a video not having tags. More work needs to be done in
this regard to make things a bit more robust, but for the time being
this gets around the problem for some videos that we have that are
just for our consumption and nobody else's.