Include support for time codes in description text
The syntax for the body of the video now contains a rule for matching a
time code, in the form of ##:##, to which it applies a numeric scope so
that they will be highlighted in the description text.

Since time codes in video descriptions are links to the video at a
specific point in time, we also have an `on_modified` event listener
that will find all time codes in the description and mark them up with
a stippled underline so that it's apparent that they're links.