Implement a command to get a video link
This implements a new command that's able to gather the link to any
particular video based on a `view` that's inside of a YouTube window.

The ID for the video comes from the window properties (set when the
video window is created).

The command is also context sensitive in that if an event is provided
that points at a time code, the time code will be converted into an
appropriate query string to append to the overall URL for the video,
allowing the link to point to a specific point in the video.

Lastly, (for now) the command can either copy the URL to the clipboard,
open it in the browser, or both. As discussed near the end of the
Devember Day 1 live stream, the code in this command is ripe for being
factored out into a helper or base class of some sort, which would split
this command into two, which is overall more logical.