Include custom context menus for YouTube views
This simple change overrides the context menu used for views in YouTube
windows to have more appropriate, relevant commands. Many commands that
are in the default context menu or added there by packages are not
useful in a YouTube view.

Each of the three types of views has its own unique menu, though for
the most part they're all the same and include entries for the newly
added command to copy the video link or open the video for viewing in
a browser.

In addition, the menu for the description/body area also includes an
item that allows for inserting the Table of Contents from a Camtasia
project file, which is a command that already existed and is bound to
a command in the command palette by default.