Fix a bug in the Devember Day 1 command
This fixes a small issue in the command that was implemented as a part
of yesterday's live stream. The command made sure to enable itself only
in a YouTube window, but it actually needs to also constrain itself to
only working in a window with a video associated with it because it
needs a video ID in order to be able to generate links.