Add a new setting and clean up settings handling
Ostensibly, this adds in a new setting regarding whether or not the
(as yet to be added) log panel should automatically display itself or
not.

Along the way, settings handling was cleaned up a bit, since there were
a couple of places in which the settings were being manually loaded
instead of using the endpoint.

This also sets up the proper global defaults for the settings endpoint
as well.