Implement an output log panel
This implements a log output panel heavily influenced by the one in the
SubliNet package.

On package load we ensure that every window has an output panel in it,
and as new windows are created they get a new output panel with a clone
of the current output. Our log routines sends its output to the output
panel in all current windows as well.

This sets us up so that no matter what window we're in, there is always
an up to date log of what's happening in the package.

Our log function works similar to others I've implemented in other
packages, but here logs go to the panel (unless they're errors) by
default. The caller can also indicate if this log should make the panel
auto open.

The auto-open itself is controlled by the newly added setting, which
says whether this should always be respected, never be respected, or
respected but with an automatic close after a timeout.