Refactor code into a new module to avoid problems
We're having a few issues with dependency loops happening in the code,
where module A imports module B, which needs some imports from module
C, which needs something from module A, causing issues.

In order to get around that, the common code that needs to be shared by
lots of different places has been hoisted out into its own module in
the lib folder, which should stop the error from occurring
(fingers crossed).

The idea is that things in `lib` should probably not be importing 
anything from `code`, and anything in `lib` can safely import from other
parts as long as there's no loop in that small load.