Refactor YouTube link generation
This is the refactor that was mentioned in the Day 1 live stream. 

There are a few new routines in the new `util` library for this. One of
them can generate a link out of a bare video ID and an optional time
code (in string or number format) which is usable by anyone with that
information available.

We also have a method to get a time code string out of a `view` based on
an `event` that indicates where a mouse access happened (i.e. in a
context sensitive way). 

Lastly, there's a link generation method that can take a view and return
back a link to the video associated with it's window, optionally also
grabbing and using the time code as well.