Expand time code parsing functionality
This simple change adds some extra power to our ability to understand
and parse time codes.

Previously, we handled time code in an MM:SS format, where each part
needed to be two digits (so, leading zeros). However, this is not
entirely valid because it's possible for a code to have just a single
digit minute.

The syntax has been expanded to support HH:MM:SS, with the hour part
optional and the HH and MM supporting a single digit if desired. SS must
always be two digits, and it's expected (but not required) that the MM
will be two digits if there are hours.

The parsing code now more robustly calculates the time code for things
by splitting the code at the colons instead of assuming specific sizes
for each part.