Fix TOC insert in video descriptions
This is a simple fix to the context menu command for inserting a table
of contents from a Camtasia project file, by having the menu entry
actually tell it to insert (the default is to copy to the clipboard
instead).

I also noticed that the `sublime.open_dialog()` call did not have an
appropriate argument, which may explain why it doesn't properly filter
file types on windows. Regardless of this change it also does not 
appear to work in Linux regardless, though this may be a GTK issue.