Clarify error returns from the the YouTube API
This "fixes" error handling by having the networking code return back
the error object directly instead of coming up with just the error
message.

This allows for the recipient of the error to gather more information
than just the message itself, such as the error code and a detailed
list of errors.

Currently all that's used is the error message and code, but this opens
the API somewhat in this regard.