Flesh out the channel details request/response
Originally most of the requests used in the package come from an API
sample, and a couple of others were added to get a bit more
functionality as part of some testing.

We're starting now on actually formalizing the requests that we're
making, the results that they return back, and why they're requesting
what they do.

As a first step, this documents what's being gathered for the channel
details including what each of the request items are for.