Flesh out the video details request/response
As a last cleanup on the requests that we already have available, this
cleans up the video details request in a vein similar to the other
cleanups.

This request gets us video details including things like the number of
views, likes and dislikes.

As a next step, this request (maybe others?) should include the ability
to specify the part provided and more than one item in the list, so that
we can use this to fetch ad-hoc details if desired.