Fix an issue introduced during request refactoring
While cleaning up the YouTube API request code yesterday, I
inadvertently named some of the request objects literally "request",
which works but masks the incoming request parameters, which is not a
good idea since potential upcoming changes require being able to access
it.

This resolves that issue to make things a little clearer.