Videos don't always contain tags; we know this
This is a recurrence of something that was previously noted (prior to
Devember) but which was inadvertently forgotten during the API refactor.

One of the things the `snippet` part of a video details can return is
the list of tags associated with the video. However, not all videos have
tags, and in that case this member is not present (instead of being say
an empty array).

This means that trying to view the details on a video that has no tags
causes an error when they're trying to be fetched. The resolution here
is the same as last time, which is to use `get` instead of a direct
access, which dotty supports natively.