Optionally fetch full video details in playlists
When requesting the contents of a playlist, one of the things you can
get is details on the actual videos contained within. However, this does
not cover ALL possible video details (such as likes or views), which
then requires a separate request.

What we can do however is fetch the playlist contents with a minimum of
details, which provides just the video ID and publish date. From that we
can grab video details based on the list.

Unfortunately, though the video details request can take a list of items
it is constrained to what appears to be 50 items, which it will not page
or otherwise allow you to filter.

So, to get around that we batch out the requests in groups of 50. This
slows down the requests somewhat, so it's optional by parameter.
However anything that makes the request needs to be aware of which way
they asked because some of the data layout changes as a result.