Allow data requests to specify the request fields
In preparation for our upcoming work in altering how data caching works,
lay the foundation for what we expect out of the request API, which is
to be able to specify the data we're most interested in.

Previously making a specific request would return exact data. However,
we want to be able to return as little or as much as we want, which has
different quota costs at times and also can be faster depending on what
we're doing.

As written here, the request object can contain a `part` to be used to
gather the data, with a default being provided if none is given. The
default is what used to be hard coded, so that no external changes are
needed.