Clean up the method of grabbing the cache dir
For reasons I'm unable to adequately explain, for some reason back in
Devember 2019 when I wrote the code for caching credentials I manually
set up a path to the Cache folder by going into a relative path from
the Packages folder.

There is in fact (and has been for quite some time) a method for getting
the cache folder directly.