Implement a busy spinner class in the status bar
One of the issues that's been outstanding for over a year and always
on the radar is a busy spinner to indicate when the network thread is
busy trying to fetch data.

This commit brings in an implementation of a busy spinner that displays
information in the status bar with an animation while an operation is in
progress.

This is mostly the one from OverrideAudit, but augmented to not follow a
thread directly, and with a changed display. Also, inspiration was taken
from someone in the Sublime Discord (rwols?) in that having the spinner
be a context manager is a handy thing to do. After a bit of brief
investigation it seems that this is pretty easy to accomplish.