Use a busy spinner for network requests
Use the new busy spinner class to wrap all network requests so that we
can track lengthy ones as they're taking place.

This is done directly in the single point in the Network thread where it
dispatches requests, which is a quick way to provide coverage. It may be
required to modify this when we get the "logging in hangs the network
thread if the user doesn't succeed in logging in" problem solved, since
that may require more finagling.