Implement a reason for request info field
In our Request class, this introduces a new field that is used to
specify why the request is being made. This will be used to set the
text label of any spinner that's displayed while this operation is in
progress.

As a first step, the value of this field defaults to the request name so
that in the near term everything continues to work until we update the
code.