Enforce not being able to close a YouTube view
In a YouTube window, we have custom views for displaying the title, body
and tags of a video. These are scratch buffers that are not persisted
and thus there is no way to recover the contents of one should you
close it when you didn't mean to.

This change include a custom key binding that blocks the close key from
being active while in a YouTube view.

This leaves the command still available via the menu, but that is a more
direct action to take that is unlikely to be an accident, so we're
going to leave that alone since we may at some point want to actually
close the view.