Clean up our body syntax a little bit
This is something that I was experimenting with on the last live stream,
but I derped out and put the meta directive in the wrong place and
messed the whole thing up.

This tweaks the syntax definition for video bodies such that the first
line is still specially colored, but also supports timecode, which was
previously not possible.