Refactor the video selection quick panel code
To prepare for some upcoming changes planned for this evening, this
factors away the common code in the list videos and video details
commands that displays a quick panel with videos in it so that it's
handled in one location only.

The idea is that you provide it a list of video details and a callback,
and it will tell you which one the user selected, if any.