Fix the playlist_contents request to use part
The original intent for last changes to the playlist_contents method was
to allow you to optionally specify what video details you want when you
request, with there being a default.

However, this did not actually work as expected for the details request
and this didn't actually do anything. This has now been resolved.