Make the video selection quick panel more betterer
Using the new powers of the QuickPanelItem in ST4, this enhances the
video selection quick panel to use kind information and annotations.

We use distinct kind information based on the status of a video (public,
private or unlisted) and also display information on the view count,
likes and dislikes for each video.