Rename the view command to edit instead
The video_details command has been renamed to edit_video_details to
better indicate what it's for; although it can be used to view details,
the primary intent is to allow the editing of details eventually.

In addition, the video details that it requires are already being
gathered by the request that gathers the contents of the uploaded video
list, so an extraneous request is removed.