Rename list command to get link command instead
As in the previous change, this renames the command that used to be
called list videos (which was a holdover to what the original sample
code used) to something that actually makes sense for what the command
does now.

The name is similar to an existing command, but that one is for getting
link information from a buffer and not for asking the user to select a
video. These may be merged at some point in the future.