Include ability to link to a TOC entry in videos
This augments the command to copy a video link so that if the video
description contains a table of contents, when chosen you get asked to
select which of the table of contents entry you want to link to.

There is not currently an item here to select copying a link without a
timecode on it, but per YouTube's spec in order to have a valid table
of contents the first item in the table of contents should start at 0,
which does the same thing. This is not yet enforced in the editor,
however.