Change the criteria for video sorting
This is something that was suggested by Recchan in the last Devember
live stream; I've been using it since then and have decided that it
seems superior to sorting by the title since we generally want to find
the title by filtering on the name anyway. This way we can get an
overview of how the best videos are doing, so what the heck, why not?