Include a new select_timecode method
In a fashion similar to the previous commits, this implements a new
chooser which, when given a single video, gathers the table of contents
from the video and prompts the user via a quick panel to choose one of
the items, which is displayed with the associated table of contents
text.

The callback is given the time code selected as well as the associated
text, and also supports the ability to "go back" using the same
mechanism we was used in the select_video change previously.