Add the ability to go back in select_video
The select_video method for selecting a specific video from a list of
videos has been modified to allow for the ability to "go back" to a
prior state.

The method now have an extra parameter which allows you to specify
whether or not you would like to enable the ability to go back, which
defaults to False.

When this is True, an extra item is added to the top of the video list
allowing you to "go back".

In use, the callback that's invoked gets passed the same values as it
did previously, EXCEPT that if show_back is True and the user picks the
first item in the list (the "go back"), the video that's passed to the
callback has the special sentinel video id of "_back" to indicate that
the user took this action.

It's up to the caller to handle this result as appropriate.