Include a new select_tag method for picking tags
This adds in a new method that allows you to select a tag from a list of
tags. 

The method needs to be provided either an array of videos OR a special
tag list value retrieved from a previous call. The method will display
a list of all unique tags found amongst the videos in the list for the
user to pick.

The provided callback is given the selected tag as well as a dictionary
that contains all of the tags and the videos associated with it. From 
this the caller can determine if desired what videos have this tag.

The returned tag list can be provided in a future call to select_tag
(presuming the video list is still the same) instead of the list of 
videos, which causes select_tag to use that list instead of trying to
calculate it again.