Merge changes from experimental branch
On Devember days 9 and 10, there was an experimental branch with a
command that was a duplicate of this original command except that it
included the ability to browse videos by tags instead of just by name.

This brings those changes to the original command that spawned that one,
which otherwise had identical functionality and differs only in the
ability to choose by tag instead.

This uses features recently added like browsing for timecodes and going
back during a browse, realizing a final version of the code so that the
test branch can be discarded.