Capture arguments in the YouTubeRequest helper
This class is the base class for any command which wants to talk to
the YouTube API; it handles a lot of the low level details.

The helper has a run() method which can take any arguments, which were
previously ignored. This modification stashes the arguments that were
given to the run method so that any of the internal code can use the
arguments to determine what they'd like to do.