Augment the get link command for optional tags
This augments the changes recently made to the command to get a video
link so that instead of always using tags as the first browse step, it
can now do that optionally.

This is done by way of a new argument called "by_tags". When this is
set to true, the command behaves as it did before this modification.

When the value is False (or not provided), the behaviour reverts to the
previous behaviour, in which you see just a list of every possible
video instead of tags.

In this case the callback stores the list of all videos in the place
that the tag list would normally be stored, since those two things are
mutually exclusive for this purpose.