Include a request to gather playlists
This includes a new network request to gather the list of playlists,
either for a specific channel or for the currently logged in user.

Investigation shows that when used for your own channel, requesting
either way returns the same channels; the result is only different if
you were to ask for the playlists associated with a channel that's not
your own (in which case you only see public videos).

Of note is that the special playlists for likes, uploads and favorite
videos is not returned as a part of this list; those are only available
in the channel details.