Include the ability to select a playlist
In the vein of the recently added helper functions to allow you to
select various items, this change adds in the ability to select a
playlist from a list of playlists.

This also includes the ability to go back, to retain parity with the
other methods which now all do the same thing.