Hoist sorting from the selection routines
The helper routines for allowing you to select a playlist, tag and video
should not sort the data they're given, but should instead display it in
the order provided.

This puts the onus on the caller to sort the data (if they would like it
sorted) instead, which seems saner. The only exception to this is the
tag selection, since that is delivered as a dictionary, though a future
change could easily allow for an extra argument. This seems less wonky
than sorting the other more key data items.

This sets the stage for a future change, wherein when we select by
playlists we can sort it based on a given criteria but still ensure that
the user upload playlist is in a known location (such as the top of the
list).