Do an experimental modification to allow browsing by playlists
This provides an experimental augmentation to the main video choice
command that allows for choosing a playlist before proceeding instead
of being hardcoded to assume that you want to choose from the uploads
playlist.

This currently works in a forward direction, but fails when you choose
to go backwards to the point where you choose a new build system.

To be fully realized, the underlying network code also needs to be
adjusted to remember individual video detail queries, as it currently
caches at the upper request level. This means that if a video appears
in more than one playlist, choosing a different playlist (if that was
possible) would re-fetch the info again.

This branch, like the other experimental branch, will potentially be
redacted from the repository when the other parts of this change are
finished, depending on the state of things.