Finish implementing go back with playlists
This adds in the final changes to allow our command to go back properly
to the playlist selection page at the appropriate point, whether tags
are being used or not.

This is a relatively clean change which ends up keeping the list of
playlists (which should never change, unlike the tags or videos, which
can change based on other selections) to make the browsing easier.

More work could be done here to make the placeholders better, etc.
However before we go that way we want to either factor this code out
into a helper, or make this command a "be all".