Start implementing a new caching scheme
Currently the way that data caching is implemented, the results of very
specific queries are cached, such that changing even one of the fields
in the request causes a whole new request.

This has been fine traditionally because there were only a few queries
that always used the same values. However now we are running into a
situation in which we might want to make different queries that share
common sub-queries.

To that end, the caching mechanism needs to change. The new chosen
scheme is that the cache is stored in the network thread instead of the
manager. This puts the responsibility to store data and understand
requests in the request code, which is a better fit.

As a trial implementation, this alters the channel data request to do
this caching. We're going to slowly move queries over, getting the
network manager to not cache the results as we do so, so that everything
keeps working the same as the transition occurs.