Enhance our error handling in network requests
On Devember 4th, the code in the network handler added in the notion
that the error returned back should be the error object returned by the
YouTube API on a request failure instead of just the text of the
exception that occurred, which it did previously.

This causes failures if a non-http error (from YouTube) is thrown, since
the handler assumes that the error object has a `content` field, which
strings do not.

This change fixes that by catching regular errors and converting them
into an appropriate error structure.