Include data validation routines for requests
This includes a new data validation method for incoming requests. Any
request that has a minimum standard for required fields can use this to
ensure that the request doesn't proceed if it will fail.

This could also be done by having multiple subclasses of the Request
class that are all purpose built for their respective requests, but
that is more invasive that we want to get into for Devember, so that's
tabled until a future update.

Instead, this is implemented two ways; "required" and "any of"; both of
which should either be None or sets of the appropriate arguments.

"required" is a set of arguments that must absolutely be present in
order for  the request to proceed; if any of these are missing, an
exception is thrown that lists the missing items.

"any_of" is a set of arguments as well, but in this case the check is
that at least one of the items appears in the request arguments. This is
what you'd use for something that can work in a few different ways.