Validate arguments in all existing requests
This small change uses the new validator to ensure that all incoming
requests have the correct parameters.

As a part of this (and in preparation for future changes), the request
for fetching the list of playlists for a channel or user has been
modified to require you to specifically provide a channel ID, instead
of allowing you to optionally just get playlists for the current user.

This is a minor change since the result was the same either way.