Fix potential errors with logging out
This fixes a couple of issues with the code used to "log out" of YouTube
in YouTuberizer.

What this actually does is discard the current component that is used to
talk to the service, and then remove the stored credentials so that an
attempt at access requires re-authorization.

The original code did this in the wrong order, so in theory the code
code throw an exception trying to remove the stored credentials, which
would leave the service alone.

In practice this would not occur because the request code tracks the
overall result of operations and would force a re-login on the next
request anyway, which explicitly removes the service. Still; correcter
is betterer.

Additionally, this also fixes an issue where the data cache was not
cleared on logging out, which allows you to log out, then log in with a
different account and see someone else's information, which seems Not
Good (tm).