Make channel_details request return all channels
This change modifies the channel_details request such that instead of
always implicitly returning the first channel returned, it instead
returns all of the channels as a list.

In addition to this change, we now also trigger an error if there are no
channels, instead of just returning an empty list. This is because
there's not a lot that can be done if there are no channels, and so
having to check that at the call site each time seems like bad form.

In our case, a channel request always results in an array with a single
channel anyway, but it leaves the door open for future expansion and is
more correct.

As a consequence of this, commands that fetch any data that requires the
channel list needed to be slightly tweaked to account for the new
return value.