Use the public count for upload playlist size
The API doesn't return the list of videos that are in the uploads
playlist, but the channel information includes information on the count
of public videos.

In order to be able to display some sort of useful information on the
uploads playlist before going inside of it, this change uses the public
video count as the number of videos in the playlist.

The first time the contents of the uploads playlist are retrieved, the
actual count is used to update the list size.