Change the request name for channel listing
We're going to implement a new command to get channel details by ID,
so we are renaming the existing channel_details command to channel_list
to free that name for use in the upcoming change.

This request now does caching of the channel data and also stores and
returns all channels and not just the first one.