Add ability to look up channel information by id
This adds in a new request to get the information for a channel based
on its ID. This is implemented currently by returning cached channel 
data from the channel list command, executing a channel_list if that has
not already been done.

This could be augmented to actually make a request for channel
information by the actual ID, but for the time being we're doing this
because the assumption is that we will only ever look up our own
channels.