Add the ability to cache playlist contents
This is a not quite complete change to cache requests for the contents
of a playlist (i.e. fetching full on video details). 

Currently the list data is cached (so this works), but it still needs to
selectively cache the video details. However, that request needs to be
completed first.