Clean up issues with live stream caching
The "channel_list" portion of the cache is a list instead of a dict
because the result of the query is a straight array that we want to
return back.

In other cache sections the cache initializes with an empty structure
and then populates in and delete keys, but in channel_list it actually
stores and deletes the list as a whole.

For parity, "channel_list" is now initialized to an empty list on
startup and returned to an empty list. In a future commit we are going
to remove this one entirely and rely purely on the "channel_list" cache
instead, but this is good enough for now.

In addition, I noticed during the stream as we were wrapping up that
the playlist cache was incorrectly deleting the wrong things on refresh
(i.e. not using the playlist_id as the thing to delete) this is fixed
now as well.