Factor out video detail gathering
There is currently two places where video details can be gathered; in
the contents of playlists, and when requesting full video details. These
both need to do the same thing but with slightly different API points.

So, this change factors out a common version of the both versions (based
on code currently inline in the playlist contents query) and gives it
the ability to look up one or more videos with a custom part, doing any
paging needed, and update information into a cache dictionary while also
returning the results back to the caller.