Use new details request in playlist_contents
This puts the newly added functionality for requesting video details in
place in the playlist_contents request.

The determination has been made to request only as much video detail as
needed for display in the quick panel selection instead of all possible
video details that we might want.

For that reason, there is a new cache item for playlist_videos that is
an object keyed on video ID's that is for the specific use of this
request.

With this in place, no more redundant queries are made; fetching
contents of a playlist will only gather that information which is needed
to fill things out, so if a video appears in more than one playlist it
does not need to be re-gathered every time.