Include error tracking in the request tracker
Our existing code catches exceptions and returns them as errors, but
this does not capture the full stack, which can make some of the
errors rather hard to track.

This includes a slight extension to the code to display a stack trace
for the caught exception at the point were we return the failure back
so that we can check the console to see what happened.