Update the details editing command to use new API
Previously, the Details command was just using the information from the
playlist information to populate (which is technically all it needs).
However in order to better test, and just for completeness, it now uses
the new details request.

Along the way, this also picks up a small modification to sort the
contents of the uploads playlist, which is what the other command was
already doing.