Include a custom context menu in the log panel
Include a custom context menu for the YouTuberizer log output panel, so
we can include our own custom options in it. Primarily this is driven
for a desire to add in some custom commands here, but the default menu
is cluttered with many commands that don't actually apply to the panel.