Send copied links to the log as well as clipboard
This small change alters how we send links to the clipboard so that it
will also log to the console the fact that it did so. 

This is partially for debugging support, partially because the context
menu can open such a link, and partially because it's common (for me
any) to copy a link, then forget the title of the video or need to
fetch the link again, so having a hard copy is nice.