Implement cache flushing in the network code
Since we are about to embark upon trying to actually persist our cached
data to disk, we're going to first need the ability to get rid of cached
data so that we can force a refresh of data when things change.

This implements a new request which does this, and simply returns back
information indicating that the flush happened.