Finalize how to present our video link command
After much thought and deliberation, the decision has been made that the
underlying command should have default values for whether or not it
should present things with or without tags and with or without
playlists but there should be only two items in the command palette.

In particular, since we favour using playlists as an easier way to dial
in, we're assuming we always want that, but we might not always want to
navigate by tag.

As such the command now has two entries as it did before, but the
playlist option is always enabled (where previously it was not).

d