Implement an initial version of persistent caching
This puts the code in place required to actually persist our cached data
to disk and reload it back on a later run.

This goes through the encryption mechanism that we use for the
credential stor, since the cached data includes information from
private videos which could be considered sensitive if it leaked.

As a part of this change, the way we use dotty has been modified
(particularly in how we import it) because on reloads of the package
the way the previous mechanism worked, some places would get a whole new
copy of the module loaded while some would return a copy from a previous
load.

While both still worked the same (dotty does not change) it does alter
`isinstance` checks in a way that made us very sad for quite some time
in the stream.