Don't indent cached JSON data
Both the cached credentials and the cached request data is stored as
JSON and encrypted for security reasons. The saved JSON is indented to
be human readable, which makes the size of the JSON larger than it
needs to be.

This causes some (probably insignificant) slowdown in generating the
resulting JSON data, but also bloats the size of the data somewhat.

As an example, the size of the (encrypted) data jumps from 22kb for the
credentials and 880kb for the data to 39kb/1.3mb.

This is somewhat of a senseless waste since the data isn't human
readable anyway, and also the time taken to encrypt and decrypt the
data scales with the data size, so it also wastes some amount of time
as well.