Move loading of cache data into the network thread
The call to `_init_cache()` is in the constructor of the network thread,
which makes it trigger when the package loads. This is an issue because
it takes a discernable, non-zero amount of time to load and decrypt the
file if it's present, which can hang Sublime for the duration.

The simplest expedient for the moment is to move this to inside of the
network thread so that when it starts it loads its cache data up in its
own thread.

This delays the first request made, but at least that doesn't cause a
slowdown in the UI in general.