Include busy spinners on the cache routines
Include a busy spinner on the cache load and save operations so that
when they happen, we can tell something is happening (since they take a
while). These ones also have the new time functionality turned on so
that we can see how long things are happening (and more importantly,
when).