Defer cache initialization until first request
As an addendum to a previous commit, this further delays the init of the
cache until the first request is made, instead of when the thread
starts.

This is because if the request is to log out or flush the cache, time
will be spent loading it just to throw it away.

Authorization is included in the list to ignore because authentication
is needed to take any action, though it does not in fact touch the
cache at all.