Initialize the network manager last at startup
This is not critical, but as a safety on future things going weird for
us and wasting time, this small change ensures that the network manager
is created for us at the end of our startup setup.

The idea here is that in theory it could want to do something with
settings during startup, and if so it would fail weirdly until we
noticed what was happening.