Fix an issue with retaining cache on deauth
This adds in a missing line in the `deauthenticate` handler which
gets rid of the stored credentials, but not the stored cache data.

This was an oversight and potentially got lost prior to committing on
a prior Devember day and then redacted away during the period I had some
uncommitted changes for testing laying around.